Generate lifespans within the given range
The previous version of this code generated lifespans way out of range. After a few seconds, all active particles had lifespands in the million range and particles lived so long all their opacities were reduce to close enough to zero to be invisible, but still alive so the emitter wasn't allowed to create any new ones.